- **Router**: Internals modularized into focused submodules with no public API changes.
### Fixed (Unreleased)

- **Security**: `security/sanitize` now re-exports `generateNonce()` and `isTrustedTypesSupported()` for legacy deep imports.
